app/frontend/requirements.txt
requests
streamlit==0.88.0
pandas
pyyaml
